+++ src/main/java/no/nav/bidrag/beregn/underholdskostnad/beregning/UnderholdskostnadBeregning.java
@@ -1,5 +1,7 @@
 package no.nav.bidrag.beregn.underholdskostnad.beregning;
 
+import java.math.BigDecimal;
+import no.nav.bidrag.beregn.barnebidrag.periode.BarnebidragPeriode;
 import no.nav.bidrag.beregn.underholdskostnad.bo.BeregnUnderholdskostnadGrunnlagPeriodisert;
 import no.nav.bidrag.beregn.underholdskostnad.bo.ResultatBeregning;
 
@@ -14,7 +16,7 @@ public interface UnderholdskostnadBeregning {
   ResultatBeregning beregnMedForhoyetBarnetrygd(
       BeregnUnderholdskostnadGrunnlagPeriodisert beregnUnderholdskostnadGrunnlagPeriodisert);
 
-  Double beregnBarnetilsynMedStonad(
+  BigDecimal beregnBarnetilsynMedStonad(
       BeregnUnderholdskostnadGrunnlagPeriodisert beregnUnderholdskostnadGrunnlagPeriodisert);
 
   static UnderholdskostnadBeregning getInstance(){
