+++ src/main/kotlin/no/nav/bidrag/beregn/bpsandelunderholdskostnad/bo/periode.kt
@@ -3,11 +3,12 @@ package no.nav.bidrag.beregn.bpsandelunderholdskostnad.bo
 import no.nav.bidrag.beregn.felles.bo.Periode
 import no.nav.bidrag.beregn.felles.bo.PeriodisertGrunnlag
 import no.nav.bidrag.beregn.felles.enums.InntektType
+import java.math.BigDecimal
 
 data class InntektPeriode(
     val inntektDatoFraTil: Periode,
     val inntektType: InntektType,
-    val inntektBelop: Double) : PeriodisertGrunnlag {
+    val inntektBelop: BigDecimal) : PeriodisertGrunnlag {
   constructor(inntektPeriode: InntektPeriode)
       : this(inntektPeriode.inntektDatoFraTil.justerDatoer(),
       inntektPeriode.inntektType,
@@ -19,7 +20,7 @@ data class InntektPeriode(
 
 data class UnderholdskostnadPeriode(
     val underholdskostnadDatoFraTil: Periode,
-    val underholdskostnadBelop: Double) : PeriodisertGrunnlag {
+    val underholdskostnadBelop: BigDecimal) : PeriodisertGrunnlag {
   constructor(underholdskostnadPeriode: UnderholdskostnadPeriode)
       : this(underholdskostnadPeriode.underholdskostnadDatoFraTil.justerDatoer(),
       underholdskostnadPeriode.underholdskostnadBelop)
